+++ src/main/java/com/nemonotfound/entity/mob/NecromancerEntity.java
@@ -1,24 +1,24 @@
 package com.nemonotfound.entity.mob;
 
+import com.nemonotfound.entity.goal.NecromancersSummonGoal;
 import com.nemonotfound.item.ModItems;
 import com.nemonotfound.sounds.ModSounds;
 import net.minecraft.entity.EntityData;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.SpawnReason;
-import net.minecraft.entity.ai.goal.*;
+import net.minecraft.entity.attribute.DefaultAttributeContainer;
 import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.damage.DamageSource;
 import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.entity.effect.StatusEffects;
 import net.minecraft.entity.mob.AbstractSkeletonEntity;
-import net.minecraft.entity.passive.IronGolemEntity;
-import net.minecraft.entity.passive.WolfEntity;
-import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.projectile.ArrowEntity;
 import net.minecraft.entity.projectile.PersistentProjectileEntity;
 import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.NbtCompound;
 import net.minecraft.sound.SoundEvent;
+import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.LocalDifficulty;
 import net.minecraft.world.ServerWorldAccess;
@@ -29,8 +29,29 @@ import java.util.Objects;
 
 public class NecromancerEntity extends AbstractSkeletonEntity {
 
+    private int summoningTicks;
+
     public NecromancerEntity(EntityType<? extends AbstractSkeletonEntity> entityType, World world) {
         super(entityType, world);
+    }
+
+    @Override
+    public void readCustomDataFromNbt(NbtCompound nbt) {
+        super.readCustomDataFromNbt(nbt);
+        this.summoningTicks = nbt.getInt("SummoningTicks");
+    }
+
+    @Override
+    public void writeCustomDataToNbt(NbtCompound nbt) {
+        super.writeCustomDataToNbt(nbt);
+        nbt.putInt("SummoningTicks", this.summoningTicks);
+    }
+
+    public static DefaultAttributeContainer.Builder createNecromancerAttributes() {
+        return AbstractSkeletonEntity.createAbstractSkeletonAttributes()
+                .add(EntityAttributes.GENERIC_MAX_HEALTH, 100)
+                .add(EntityAttributes.GENERIC_ARMOR, 10)
+                .add(EntityAttributes.GENERIC_ATTACK_DAMAGE, 4);
     }
 
     @Override
@@ -58,6 +79,14 @@ public class NecromancerEntity extends AbstractSkeletonEntity {
         return ModSounds.ENTITY_NECROMANCER_STEP;
     }
 
+    public SoundEvent getSummonPrepareSound() {
+        return SoundEvents.ENTITY_EVOKER_PREPARE_SUMMON;
+    }
+
+    public SoundEvent getSummonSound() {
+        return SoundEvents.ENTITY_EVOKER_CAST_SPELL;
+    }
+
     @Override
     public float getSoundVolume() {
         return 0.5F;
@@ -65,15 +94,8 @@ public class NecromancerEntity extends AbstractSkeletonEntity {
 
     @Override
     protected void initGoals() {
-        this.goalSelector.add(2, new AvoidSunlightGoal(this));
-        this.goalSelector.add(3, new EscapeSunlightGoal(this, 1.0));
-        this.goalSelector.add(3, new FleeEntityGoal<>(this, WolfEntity.class, 6.0f, 1.0, 1.2));
-        this.goalSelector.add(5, new WanderAroundFarGoal(this, 1.0));
-        this.goalSelector.add(6, new LookAtEntityGoal(this, PlayerEntity.class, 8.0f));
-        this.goalSelector.add(6, new LookAroundGoal(this));
-        this.targetSelector.add(1, new RevengeGoal(this));
-        this.targetSelector.add(2, new ActiveTargetGoal<>(this, PlayerEntity.class, true));
-        this.targetSelector.add(3, new ActiveTargetGoal<>(this, IronGolemEntity.class, true));
+        this.goalSelector.add(4, new NecromancersSummonGoal(this));
+        super.initGoals();
     }
 
     @Override
@@ -89,10 +111,23 @@ public class NecromancerEntity extends AbstractSkeletonEntity {
     @Nullable
     public EntityData initialize(ServerWorldAccess world, LocalDifficulty difficulty, SpawnReason spawnReason, @Nullable EntityData entityData) {
         EntityData initializedEntityData = super.initialize(world, difficulty, spawnReason, entityData);
-        Objects.requireNonNull(this.getAttributeInstance(EntityAttributes.GENERIC_ATTACK_DAMAGE)).setBaseValue(4.0);
+        Objects.requireNonNull(this.getAttributeInstance(EntityAttributes.GENERIC_ATTACK_DAMAGE)).setBaseValue(10.0);
         this.updateAttackType();
 
         return initializedEntityData;
+    }
+
+    //TODO: change attack type to necromancersScepter
+    public void updateAttackType() {
+        super.updateAttackType();
+    }
+
+    @Override
+    protected void mobTick() {
+        super.mobTick();
+        if (this.summoningTicks > 0) {
+            --this.summoningTicks;
+        }
     }
 
     @Override
@@ -102,5 +137,9 @@ public class NecromancerEntity extends AbstractSkeletonEntity {
             ((ArrowEntity)persistentProjectileEntity).addEffect(new StatusEffectInstance(StatusEffects.HUNGER, 100));
         }
         return persistentProjectileEntity;
+    }
+
+    public void setSummoningTicks(int summoningTicks) {
+        this.summoningTicks = summoningTicks;
     }
 }